+++ examples/utility/rad.h
@@ -4,5 +4,5 @@
 	#define M_PI 3.14159265358979323846
 #endif
 
-#define RAD(x) ((double) (x) / 180.) * M_PI
-#define DEG(x) ((double) (x) / M_PI) * 180.
+#define RAD(x) ((float) (x) / 180.) * M_PI
+#define DEG(x) ((float) (x) / M_PI) * 180.
